Extract build-only TypeScript configuration
Currently, we exclude test files from TypeScript's purview, so that only
files in `src/` get emitted to `dist/`. This makes sense from a release
perspective, but it interferes with development.

One example of this interference involves module augmentation. We have a
file `dependencies.d.ts` which we can use to augment types for various
dependencies. Right now we don't have much in this file, but in a future
commit we will be backfilling a lot of types using this mechanism.
TypeScript will see and incorporate these module augmentations for
source files, but not for test files. This causes in-editor
inconsistencies with type analysis and, in some cases, warnings such as:

* `import` can only be used with `esInteropModuleFlag: true`
* `X` is missing type definitions

To fix this, this commit breaks off the build-specific configuration
from `tsconfig.json`. `tsconfig.json` will still be utilized by VSCode
or whatever editor is being used and contains the widest possible
settings; `tsconfig.build.json` narrows the set of emitted files as in
the original configuration.